+++ src/singletons/Settings.hpp
@@ -690,6 +690,11 @@ public:
 
     QStringSetting additionalExtensionIDs{"/misc/additionalExtensionIDs", ""};
 
+    BoolSetting showDeleteMessageOptionInMessageMenu = {
+        "/moderation/showDeleteMessageOptionInMessageMenu",
+        true,
+    };
+
 private:
     ChatterinoSetting<std::vector<HighlightPhrase>> highlightedMessagesSetting =
         {"/highlighting/highlights"};